Documented InputDevice

+++ src/InputDevice/InputDevice.cpp
@@ -1,8 +1,7 @@
-#include <cstdint>
 #include <initializer_list>
 #include "InputDevice.hpp"
 
 namespace Intel8080 {
-    InputDevice::InputDevice(const std::initializer_list<uint8_t>& inputPortNumbers)
+    InputDevice::InputDevice(const std::initializer_list<int>& inputPortNumbers)
         : inputPortNumbers{inputPortNumbers}{}
 }
