+++ src/OutputDevice/OutputDevice.cpp
@@ -1,8 +1,7 @@
-#include <cstdint>
 #include <initializer_list>
 #include "OutputDevice.hpp"
 
 namespace Intel8080 {
-    OutputDevice::OutputDevice(const std::initializer_list<uint8_t>& outputPortNumbers)
+    OutputDevice::OutputDevice(const std::initializer_list<int>& outputPortNumbers)
         : outputPortNumbers{outputPortNumbers}{}
 }
